implemented so a post can be answered.
Groups created if an assisting user answers, and no group does not already exist.

User is included in existing group if existing.

Made some modifications to some structs in order to simplify, and correct a mistake (models.User using models.UserResponse, which should not be used there, only in models.Post)